+++ plugins/Moodle/functions.php
@@ -491,10 +491,10 @@ function MoodleTriggered( $hook_tag, $arg1 = '' )
 				Moodle( $modname, 'core_course_update_courses' );
 
 				//update teacher too
-				global $columns, $current;
+				global $columns, $current_cp;
 
 				if ( ! empty( $columns['TEACHER_ID'] )
-					&& $columns['TEACHER_ID'] != $current[1]['TEACHER_ID'] )
+					&& $columns['TEACHER_ID'] != $current_cp[1]['TEACHER_ID'] )
 				{
 					Moodle( $modname, 'core_role_unassign_roles' );
 					Moodle( $modname, 'core_role_assign_roles' );